MNT: rewire uv workspace with inifix as the root
