--- lib/inifix/pyproject.toml
@@ -1,64 +1,0 @@
-[build-system]
-requires = ["hatchling", "hatch-vcs"]
-build-backend = "hatchling.build"
-
-[project]
-name = "inifix"
-version = "6.1.2"
-description = "An I/O library for Pluto-style ini files."
-authors = [
-    { name = "C.M.T. Robert" },
-]
-license = "GPL-3.0-only"
-license-files = ["LICENSE"]
-classifiers = [
-    "Intended Audience :: Developers",
-    "Intended Audience :: Science/Research",
-    "Development Status :: 5 - Production/Stable",
-    "Operating System :: POSIX :: Linux",
-    "Operating System :: MacOS",
-    "Operating System :: Microsoft :: Windows",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Programming Language :: Python :: 3.14",
-    "Programming Language :: Python :: 3.15",
-    "Programming Language :: Python :: Free Threading :: 3 - Stable",
-    "Topic :: Scientific/Engineering",
-    "Typing :: Typed",
-]
-requires-python = ">=3.11"
-dependencies = []
-
-[project.readme]
-file = "README.md"
-content-type = "text/markdown"
-
-[project.urls]
-Homepage = "https://github.com/neutrinoceros/inifix"
-Changelog = "https://github.com/neutrinoceros/inifix/blob/main/CHANGELOG.md"
-
-[tool.hatch]
-version.source = "vcs"
-
-[tool.hatch.build.targets.sdist]
-only-include = [
-    "src",
-    "tests",
-    "LICENSE",
-    "CHANGELOG.md",
-    "README.md",
-]
-
-[tool.hatch.build.hooks.vcs]
-version-file = "src/inifix/_version.py"
-template = '''
-# This file is automatically generated by Hatch
-from typing import Final
-
-__all__ = ["__version__", "__version_tuple__"]
-
-__version__: Final = {version!r}
-__version_tuple__: tuple[int, int, int] = {version_tuple!r}
-'''
