+++ pyproject.toml
@@ -1,26 +1,67 @@
 [build-system]
-requires = ["flit_core >=3.2,<4"]
-build-backend = "flit_core.buildapi"
+requires = ["hatchling", "hatch-vcs"]
+build-backend = "hatchling.build"
 
 [project]
-name = "inifix-cli"
-version = "0.1.0"
-description = "A command line interface for inifix."
+name = "inifix"
+version = "6.1.2"
+description = "An I/O library for Pluto-style ini files."
 authors = [
     { name = "C.M.T. Robert" },
 ]
+license = "GPL-3.0-only"
+license-files = ["LICENSE"]
+classifiers = [
+    "Intended Audience :: Developers",
+    "Intended Audience :: Science/Research",
+    "Development Status :: 5 - Production/Stable",
+    "Operating System :: POSIX :: Linux",
+    "Operating System :: MacOS",
+    "Operating System :: Microsoft :: Windows",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
+    "Programming Language :: Python :: 3.15",
+    "Programming Language :: Python :: Free Threading :: 3 - Stable",
+    "Topic :: Scientific/Engineering",
+    "Typing :: Typed",
+]
 requires-python = ">=3.11"
-dependencies = [
-    "click==8.3.3",
-    "inifix==6.1.2",
-    "typing-extensions>=4.4.0 ; python_full_version < '3.12'",
+dependencies = []
+
+[project.readme]
+file = "README.md"
+content-type = "text/markdown"
+
+[project.urls]
+Homepage = "https://github.com/neutrinoceros/inifix"
+Changelog = "https://github.com/neutrinoceros/inifix/blob/main/CHANGELOG.md"
+
+[tool.hatch]
+version.source = "vcs"
+
+[tool.hatch.build.targets.sdist]
+only-include = [
+    "src",
+    "tests",
+    "LICENSE",
+    "CHANGELOG.md",
+    "README.md",
 ]
 
-[project.license]
-text = "GPL-3.0"
+[tool.hatch.build.hooks.vcs]
+version-file = "src/inifix/_version.py"
+template = '''
+# This file is automatically generated by Hatch
+from typing import Final
 
-[project.scripts]
-inifix = "inifix_cli:app"
+__all__ = ["__version__", "__version_tuple__"]
+
+__version__: Final = {version!r}
+__version_tuple__: tuple[int, int, int] = {version_tuple!r}
+'''
 
 [dependency-groups]
 test = [
@@ -44,53 +85,11 @@ typecheck = [
 ]
 
 [tool.uv.workspace]
-members = ["lib/inifix"]
+members = ["cli/inifix-cli"]
 
 [tool.uv.sources]
 inifix = { workspace = true }
-
-[tool.ruff.lint]
-exclude = ["*__init__.py"]
-ignore = ["E501"]
-select = [
-    "E",
-    "F",
-    "W",
-    "C4",  # flake8-comprehensions
-    "B",   # flake8-bugbear
-    "YTT", # flake8-2020
-    "I",   # isort
-    "UP",  # pyupgrade
-    "RUF022" # unsorted-dunder-all
-]
-
-[tool.ruff.lint.isort]
-known-first-party = ["inifix"]
-
-[tool.coverage.run]
-branch = true
-source = [
-    "src/inifix_cli",
-    "tests",
-    "lib/inifix/src",
-    "lib/inifix/tests",
-]
-omit = [
-    "src/inifix_cli/__main__.py",
-    "scripts/*",
-]
-
-[tool.coverage.report]
-show_missing = true
-skip_covered = false
-
-# use exclude_also to *extend* the default exclude_lines,
-# and allow purely defensive code patterns
-exclude_also = [
-    "^\\s*raise NotImplementedError\\b",
-    "^\\s*raise RuntimeError\\b",
-    "^\\s*case _ as unreachable:",
-]
+inifix-cli = { workspace = true }
 
 [tool.pyright]
 reportImplicitStringConcatenation = false
@@ -117,3 +116,46 @@ addopts = ["-ra"]
 filterwarnings = [
     "error",
 ]
+
+[tool.coverage.run]
+branch = true
+source = [
+    "src/inifix",
+    "tests",
+    "cli/inifix-cli/src",
+    "cli/inifix-cli/tests",
+]
+omit = [
+    "cli/inifix-cli/src/inifix_cli/__main__.py",
+    "scripts/*",
+]
+
+[tool.coverage.report]
+show_missing = true
+skip_covered = false
+
+# use exclude_also to *extend* the default exclude_lines,
+# and allow purely defensive code patterns
+exclude_also = [
+    "^\\s*raise NotImplementedError\\b",
+    "^\\s*raise RuntimeError\\b",
+    "^\\s*case _ as unreachable:",
+]
+
+[tool.ruff.lint]
+exclude = ["*__init__.py"]
+ignore = ["E501"]
+select = [
+    "E",
+    "F",
+    "W",
+    "C4",  # flake8-comprehensions
+    "B",   # flake8-bugbear
+    "YTT", # flake8-2020
+    "I",   # isort
+    "UP",  # pyupgrade
+    "RUF022" # unsorted-dunder-all
+]
+
+[tool.ruff.lint.isort]
+known-first-party = ["inifix"]
