+++ src/inifix/_floatencoder.py
