+++ src/inifix/_format.py
