+++ src/inifix/_validation.py
