+++ tests/data/fargo_planet.cfg
