+++ tests/data/idefix-khi.ini
